pyproject.toml
[tool.poetry]
name = "diffusion"
version = "0.1.0"
description = ""
authors = ["JingYu Ning <115336923+jrunkening@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
torch = { url = "https://download.pytorch.org/whl/cu118/torch-2.0.1%2Bcu118-cp311-cp311-linux_x86_64.whl"}
torchvision = { url = "https://download.pytorch.org/whl/cu118/torchvision-0.15.2%2Bcu118-cp311-cp311-linux_x86_64.whl" }
pillow = "^10.0.0"
neuraloperator = "^0.2.0"
tensorly = "^0.8.1"
opt-einsum = "^3.3.0"
tensorly-torch = "^0.4.0"
wandb = "^0.15.5"
h5py = "^3.9.0"
zarr = "^2.15.0"

[tool.poetry.group.dev.dependencies]
jupyterlab = "^4.0.2"
matplotlib = "^3.7.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
